refactor: style 관련 상수 분리
