+++ frontend/src/components/FilterCategoryItem/FilterCategoryItem.tsx
@@ -9,7 +9,8 @@ import { userState } from '@/recoil/atoms';
 
 import Button from '@/components/@common/Button/Button';
 
-import { CACHE_KEY, PALETTE } from '@/constants';
+import { CACHE_KEY } from '@/constants';
+import { PALETTE } from '@/constants/style';
 
 import subscriptionApi from '@/api/subscription';
