+++ frontend/src/constants/index.ts
@@ -36,33 +36,6 @@ const STORAGE_KEY = {
   ACCESS_TOKEN: 'accessToken',
 };
 
-const PALETTE = [
-  '#AD1457',
-  '#D81B60',
-  '#D50000',
-  '#E67C73',
-  '#F4511E',
-  '#EF6C00',
-  '#F09300',
-  '#F6BF26',
-  '#E4C441',
-  '#C0CA33',
-  '#7CB342',
-  '#33B679',
-  '#0B8043',
-  '#009688',
-  '#039BE5',
-  '#4285F4',
-  '#3F51B5',
-  '#7986CB',
-  '#B39DDB',
-  '#9E69AF',
-  '#8E24AA',
-  '#795548',
-  '#616161',
-  '#A79B8E',
-];
-
 const PATH = {
   MAIN: '/',
   AUTH: '/oauth',
@@ -78,6 +51,5 @@ export {
   DAYS,
   SELECTOR_KEY,
   STORAGE_KEY,
-  PALETTE,
   PATH,
 };
