+++ frontend/src/constants/style.ts
@@ -1,0 +1,28 @@
+const PALETTE = [
+  '#FABEC0',
+  '#F85C70',
+  '#F37970',
+  '#E43D40',
+  '#F9D876',
+  '#FBE39D',
+  '#FEDA15',
+  '#CBAE11',
+  '#3D550C',
+  '#81B622',
+  '#ECF87F',
+  '#59981A',
+  '#145DA0',
+  '#0C2D48',
+  '#2E8BC0',
+  '#B1D4E0',
+  '#BEAFC2',
+  '#695E93',
+  '#8155BA',
+  '#E43480',
+  '#EAD4C0',
+  '#BBC4C2',
+  '#464033',
+  '#7E7C73',
+];
+
+export { PALETTE };
